Fix color_extensions.dart by properly casting num to int

+++ lib/utils/color_extensions.dart
@@ -11,10 +11,10 @@ extension ColorExtensions on Color {
   
   /// Bileşen değerlerini kullanarak renk oluşturur
   Color withRGBValues({int? red, int? green, int? blue, double? opacity}) {
-    final int r = red ?? this.r;
-    final int g = green ?? this.g;
-    final int b = blue ?? this.b;
-    final int a = opacity != null ? (opacity * 255).round() : this.a;
+    final int r = red ?? this.r.toInt();
+    final int g = green ?? this.g.toInt();
+    final int b = blue ?? this.b.toInt();
+    final int a = opacity != null ? (opacity * 255).round() : this.a.toInt();
     
     return Color.fromARGB(a, r, g, b);
   }
